Change the way data is written to the temporary ts segment. (#162)
* Change the way data is written to the temporary ts segment.

If I write the original data fetched from fetchFile it won't read the ts segment to combine into mp4, so I split the image and ts segment. Then write the ts segment and it works.

* Define IEND_IMAGE for MP4 conversion process

Add IEND_IMAGE constant for MP4 conversion.

* Reorganize IEND_IMAGE constant position

Moved the IEND_IMAGE constant to the end of the file for better organization.

* Update src/logic/convert-hls-to-mp4.ts

Co-authored-by: qodo-merge-pro[bot] <151058649+qodo-merge-pro[bot]@users.noreply.github.com>

* typo

* rollback set segment.uri for ffmeg

---------

Co-authored-by: Tachibana Shin <shin@shin.is-a.dev>
Co-authored-by: qodo-merge-pro[bot] <151058649+qodo-merge-pro[bot]@users.noreply.github.com>